ingestor/app/airflow_client/pyproject.toml
[tool.poetry]
name = "airflow-api-stable-client"
version = "2.10.3"
description = "A client library for accessing Airflow API (Stable)"
authors = []
readme = "README.md"
packages = [
    {include = "airflow_api_stable_client"},
]
include = ["CHANGELOG.md", "airflow_api_stable_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
